Extract paths in webpack config to paths module
- Splitting out smaller chunks makes it more maintainable.
- Naming it paths now. The only alternative name I can think of is directory.
  Perhaps this is an indication that this might be the wrong abstraction (the
  fact that it is hard to name). However, I'm keeping it as is. We will refactor
  in the future if we see fit.
